Move booking page objects to their own folder
